refactor(cli): move review render helpers to review_render.go (PRD #1009 M1)
